moved rlp calculation, fixed naming
